Update more colors. Added in some project photos
After getting through more of the fundamentals of design by Code
School I have one again gone head and updated the colors and now have
comments at the top of the css page to give me a base color so I can
build the complementary colors from there.

I also have just linked the photo gallery project to the git hub account
which I should change once I get the portfolio up on the web to put
the photo gallery up as well so that I can link to it that way instead.